add users

+++ Planteer/plants/views.py
@@ -7,6 +7,8 @@ from .forms import CommentForm
 from django.db.models import Q, Count ,Avg 
 from django.core.paginator import Paginator
 from django.contrib import messages
+from django.contrib.auth import authenticate
+
 # Create your views here.
 
 def all_plants_view(request: HttpRequest):
@@ -66,42 +68,58 @@ def details_view(request:HttpRequest, plant_id):
         if comment_form.is_valid():
             form = comment_form.save(commit=False)
             form.plant_id = plant
+            form.user = request.user
             form.save()
         else:
             print(comment_form.errors)
 
-    return render(request,'plants/details.html', {'selected_plant': plant, 'plants': recom_plants,'comments': comments})
+    return render(request,'plants/details.html', {'selected_plant': plant, 'plants': recom_plants,'comments': comments, 'user': request.user})
 
 def add_plant(request:HttpRequest):
-    countries = Country.objects.all().order_by('name')
-    if request.method == 'POST':
-        
-        plant_form = PlantForm(request.POST, request.FILES)
-        if plant_form.is_valid():
-            plant_form.save()
-            messages.success(request,'Your plant has been added')
-        else:
-            messages.error(request, "Something goes wrong")
-            return render(request, 'plants/add_plant.html', {
-                'plant_form': plant_form,
-                'categories': Plant.CategoryChoices.choices,
-                'countries': countries
-            })
+        if request.user.is_superuser:
+            countries = Country.objects.all().order_by('name')
+            if request.method == 'POST':
+                
+                plant_form = PlantForm(request.POST, request.FILES)
+                if plant_form.is_valid():
+                    plant_form.save()
+                    messages.success(request,'Your plant has been added')
+                else:
+                    messages.error(request, "Something goes wrong")
+                    return render(request, 'plants/add_plant.html', {
+                        'plant_form': plant_form,
+                        'categories': Plant.CategoryChoices.choices,
+                        'countries': countries,
+                        'user': request.user
+                    })
+                return redirect('main:home_view')
+            return render(request, 'plants/add_plant.html',{'categories': Plant.CategoryChoices.choices, 'countries': countries})
         return redirect('main:home_view')
-    return render(request, 'plants/add_plant.html',{'categories': Plant.CategoryChoices.choices, 'countries': countries})
 
 def update_plant(request: HttpRequest, plant_id):
-    plant = Plant.objects.get(pk=plant_id)
-    countries = Country.objects.all().order_by('name')
-    if request.method == 'POST':
-        plant_form = PlantForm(request.POST, request.FILES, instance=plant)
-        selected_country_ids = request.POST.getlist('countries')
-        if plant_form.is_valid():
-            plant_form.save()
-            messages.success(request,'Your plant has been updated')
-            return redirect('plants:details_view', plant_id = plant_id)
-        else:
-            messages.error(request, 'Something goes wrong')
+    if request.user.is_superuser:
+        plant = Plant.objects.get(pk=plant_id)
+        countries = Country.objects.all().order_by('name')
+        if request.method == 'POST':
+            plant_form = PlantForm(request.POST, request.FILES, instance=plant)
+            selected_country_ids = request.POST.getlist('countries')
+            if plant_form.is_valid():
+                plant_form.save()
+                messages.success(request,'Your plant has been updated')
+                return redirect('plants:details_view', plant_id = plant_id)
+            else:
+                messages.error(request, 'Something goes wrong')
+
+            return render(request, 'plants/update_plant.html', {
+                'plant': plant,
+                'plant_form': plant_form,
+                'categories': Plant.CategoryChoices.choices,
+                'countries': countries,
+                'plant_country_ids': list(plant.countries.values_list('id', flat=True)),
+                'selected_country_ids': selected_country_ids,
+            })
+
+        plant_form = PlantForm(instance=plant)
 
         return render(request, 'plants/update_plant.html', {
             'plant': plant,
@@ -109,30 +127,21 @@ def update_plant(request: HttpRequest, plant_id):
             'categories': Plant.CategoryChoices.choices,
             'countries': countries,
             'plant_country_ids': list(plant.countries.values_list('id', flat=True)),
-            'selected_country_ids': selected_country_ids,
+            'selected_country_ids': [],
         })
-
-    plant_form = PlantForm(instance=plant)
-
-    return render(request, 'plants/update_plant.html', {
-        'plant': plant,
-        'plant_form': plant_form,
-        'categories': Plant.CategoryChoices.choices,
-        'countries': countries,
-        'plant_country_ids': list(plant.countries.values_list('id', flat=True)),
-        'selected_country_ids': [],
-    })
+    return redirect('main:home_view')
 
 def delete_plant(request:HttpRequest, plant_id):
+    if request.user.is_superuser:
+        try:
+            plant = Plant.objects.get(pk = plant_id)
+            plant.delete()
+            messages.success(request, "The plant has benn deleted")
+        except:
+            messages.error(request, "Something goes wrong")
 
-    try:
-        plant = Plant.objects.get(pk = plant_id)
-        plant.delete()
-        messages.success(request, "The plant has benn deleted")
-    except:
-        messages.error(request, "Something goes wrong")
-
-    return redirect('plants:all_plants_view')
+        return redirect('plants:all_plants_view')
+    return redirect('main:home_view')
 
 def search_plant(request: HttpRequest):
     query = request.GET.get('q', '').strip()
@@ -155,7 +164,7 @@ def country_detail(request:HttpRequest, country_id):
 
     country = Country.objects.get(pk=country_id)
     country_plants = country.plant_set.all()
-    print(country_plants)
+
     return render(request, 'plants/country_detail.html',{'plants': country_plants, 'country': country})
